+++ src/components/Events/ProposalRelatedEvent.tsx
@@ -1,24 +1,11 @@
 import useFormatMessage from '../../hooks/useFormatMessage'
-import { ActivityTickerEvent, EventType, ProposalCommentedEventData } from '../../shared/types/events'
+import { ActivityTickerEvent, ProposalCommentedEventData } from '../../shared/types/events'
 import Time from '../../utils/date/Time'
-import locations from '../../utils/locations'
 import Link from '../Common/Typography/Link'
 import Markdown from '../Common/Typography/Markdown'
 import Text from '../Common/Typography/Text'
 
-export const getLink = (event: ActivityTickerEvent) => {
-  if (
-    event.event_type === EventType.ProposalCreated ||
-    event.event_type === EventType.Voted ||
-    event.event_type === EventType.ProposalCommented
-  ) {
-    return locations.proposal(event.event_data.proposal_id)
-  }
-
-  if (event.event_type === EventType.UpdateCreated || event.event_type === EventType.ProjectUpdateCommented) {
-    return locations.update(event.event_data.update_id)
-  }
-}
+import { getLink } from './utils'
 
 export default function ProposalRelatedEvent({ event }: { event: ActivityTickerEvent }) {
   const t = useFormatMessage()
